+++ server/index.js
@@ -19,7 +19,8 @@ Hull.logger.transports.console.json = true;
 
 const options = {
   hostSecret: SECRET || "1234",
-  port: PORT || 8082
+  port: PORT || 8082,
+  timeout: "2m"
 };
 
 const app = express();